Separate tracking, other fixes
