+++ build-system/dep-check-config.js
@@ -307,9 +307,9 @@ exports.rules = [
       // Accessing extension-location.calculateExtensionScriptUrl().
       'extensions/amp-script/0.1/amp-script.js->' +
             'src/service/extension-location.js',
-      'extensions/amp-access-scroll/0.1/scroll-impl.js->' +
+      'extensions/amp-access-scroll/0.1/read-depth-tracker.js->' +
           'src/service/position-observer/position-observer-impl.js',
-      'extensions/amp-access-scroll/0.1/scroll-impl.js->' +
+      'extensions/amp-access-scroll/0.1/read-depth-tracker.js->' +
           'src/service/position-observer/position-observer-worker.js',
 
     ],
